:sparkles: text display, character name input :sparkles:
The story text for the game now has a placeholder. An array of sentences is entered in unity, the characters of each sentence are displayed individually and a continue button starts the next sentence display. There is also an input name feature on the menu screen which saves this into the "GameSettings" script to be called back later. (also a couple of animations added to the text).